% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCorr.R
\name{addCorr}
\alias{addCorr}
\title{Add correlation matrix to TriPlotObject}
\usage{
addCorr(TPObject, Corr)
}
\arguments{
\item{TPObject}{A TriPlotObject}

\item{Corr}{A correlation matrix; TPO components in columns and correlation variables in rows. Make sure that `Corr` has rownames.}
}
\value{
A TriPlotObject with added correlations
}
\description{
Add correlation matrix to TriPlotObject
}
\examples{
See example under triPlot()


}
