% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addRisk.R
\name{addRisk}
\alias{addRisk}
\title{Add risk estimates to TriPlotObject}
\usage{
addRisk(TPObject, Risk)
}
\arguments{
\item{TPObject}{A TriPlotObject}

\item{Risk}{A list of risk risk estimation matrix with components in rows and estimates, margin-of-error and (optionally) p-values in columns. The margin-of-error corresponds to half the width of the confidence interval (i.e. z * se).}
}
\value{
A TriPlotObject with added risk estimates
}
\description{
Add risk estimates to TriPlotObject
}
\examples{
See example under triPlot()
}
