% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTPO.R
\name{checkTPO}
\alias{checkTPO}
\title{Provide detailed information about TriPlotObject}
\usage{
checkTPO(
  TPObject,
  heatmap = TRUE,
  colorScheme = "MO",
  scaleRisk = TRUE,
  cluster = "none"
)
}
\arguments{
\item{TPObject}{A TriPlotObject}

\item{heatmap}{Whether to plot heatmap with correlations and risks (defaults to TRUE)}

\item{colorScheme}{Which color scheme to use for heatmap: 'MO' (mixOmics original; default) or high-contrast 'BWR' (Blue/white/red)}

\item{scaleRisk}{Whether to scale value of risks to max of correlations (defaults to TRUE). This makes risk estimate scale comparable to correlations in the heatmap.}

\item{cluster}{Whether to cluster x and/or y axes (defaults to 'none'). See ?mixOmics::cim for details.}
}
\value{
Detailed information and heatmap of correlations and risks in TriPlotObject
}
\description{
Provide detailed information about TriPlotObject
}
\examples{
See example under triPlot()
}
