% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficient_OR.R
\name{coefficient_OR}
\alias{coefficient_OR}
\title{Convenience function to calculate crude risk estimates using conditional logistic regression}
\usage{
coefficient_OR(TPObject, caseStat, pair, CI = 0.95)
}
\arguments{
\item{TPObject}{A TriPlotObject}

\item{caseStat}{Vector with case/control status of TPO observations (score rows)}

\item{pair}{Vector with case/control pairs of TPO observations (score rows)}

\item{CI}{Confidence Interval for the risk estimate (defaults to 0.95)}
}
\value{
A risk estimation matrix with components in rows and (estimates, margin-of-error and p-values) in columns. The margin-of-error corresponds to half the width of the confidence interval (i.e. z * se).
}
\description{
Convenience function to calculate crude risk estimates using conditional logistic regression
}
\examples{
See example under triPlot()
}
