% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force_parttype.R
\name{force_parttype}
\alias{force_parttype}
\title{To make a dataframe all numeric or all factor,or all character}
\usage{
force_parttype(
  Dataframe,
  transform_part = c("numeric", "factor", "character"),
  variables_names
)
}
\arguments{
\item{Dataframe}{A data frame}

\item{transform_part}{"numeric","factor","character"}

\item{variable_names}{The variable names that you want to do the transformation}
}
\value{
A data frame that is all numeric
}
\description{
To make a dataframe all numeric or all factor,or all character
}
\examples{
See example under triPlot()
}
