% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force_type.R
\name{force_type}
\alias{force_type}
\title{To make a dataframe all numeric or all factor,or all character}
\usage{
force_type(Dataframe, transform_all = c("numeric", "factor", "character"))
}
\arguments{
\item{Dataframe}{A data frame}

\item{transform_all}{"numeric","factor","character"}
}
\value{
A data frame that is all numeric
}
\description{
To make a dataframe all numeric or all factor,or all character
}
\examples{
See example under triPlot()
}
