% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCorr.R
\name{makeCorr}
\alias{makeCorr}
\title{Convenience function for normal correlation of data with TriPlotObject scores}
\usage{
makeCorr(TPObject, corrData, use = "pairwise", method = "spearman")
}
\arguments{
\item{TPObject}{A TriPlotObject}

\item{corrData}{Data to correlate to the TPO scores; observations in rows and variables in columns. Ensure that observations (rows) match between TPO scores and correlation data.}

\item{use}{See ?cor documentation (defaults to 'pairwise'). An optional character string giving a method for computing covariances in the presence of missing values.
This must be (an abbreviation of) one of the strings "everything", "all.obs", "complete.obs", "na.or.complete", or "pairwise.complete.obs".}

\item{method}{"pearson"(default),"spearman" or "kendall"}
}
\value{
A correlation matrix
}
\description{
Convenience function for normal correlation of data with TriPlotObject scores
}
\examples{
See example under triPlot()

}
