% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTPO.R
\name{makeTPO}
\alias{makeTPO}
\title{Prepare TriPlotObject which memorize the matrix of scores, the matrix of loadings, number of observations and variables,
number of components, number of associated correlations, number of associated risks}
\usage{
makeTPO(scores, loadings, compLimit, hide_question = F)
}
\arguments{
\item{scores}{Scores have observations in rows and components (or factors) in columns}

\item{loadings}{Loadings have variables in rows and components (or factors) in columns}

\item{compLimit}{Option limit to the number of component}

\item{hide_question}{if question is hiden or not}
}
\value{
A TriPlotObject
}
\description{
Prepare TriPlotObject which memorize the matrix of scores, the matrix of loadings, number of observations and variables,
number of components, number of associated correlations, number of associated risks
}
\examples{
See example under triPlot()
}
