% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_scan.R
\name{normal_scan}
\alias{normal_scan}
\title{Scan the data frame by each column and see which one is. transformation and}
\usage{
normal_scan(
  Dataframe,
  Group = NULL,
  figure = F,
  p_threshold = 0.05,
  perform_transform = c("none", "log_specfic", "log_all", "scale_specific", "scale_all",
    "squareroot_all", "squareroot_specific", "cubicroot_all", "cubicroot_specific")
)
}
\arguments{
\item{Dataframe}{rows as observations columns as}

\item{Group}{Group vairable if you want to separate one variable to different groups to examine normality}

\item{figure}{select figure is T only when there are small number of variables to show}

\item{p_threshold}{a p value, usually set as 0.05}

\item{perform_transform}{log, or square root ot cube root, or scale to unit}
}
\value{
A vector with adjusted partial correlations
}
\description{
Scan the data frame by each column and see which one is. transformation and
}
