% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_plots.R
\name{PCA_plots}
\alias{PCA_plots}
\title{PCA_plots}
\usage{
PCA_plots(
  dataframe,
  plottype = c("scree", "score", "loading", "scoreloading"),
  pc_type = c("prcomp", "principle"),
  pc_num = 5,
  scale = T,
  center = T,
  rotate = "none",
  size_variable = NULL,
  size_variable_name = NULL,
  color_variable = NULL,
  color_variable_name = NULL,
  shape_variable = NULL,
  shape_variable_name = NULL,
  scale_scoreloading = T,
  first_PC = 1,
  second_PC = 2
)
}
\arguments{
\item{dataframe}{row as observations, column as variables}

\item{plottype}{scree score loading 2 options}

\item{pc_type}{prcomp or principle}

\item{pc_num}{number of components}

\item{scale}{true by default}

\item{center}{true by defaule}

\item{rotate}{"none", "varimax", "quartimax", "promax", "oblimin", "simplimax", and "cluster"}
}
\value{
A PCA plot
}
\description{
PCA_plots
}
