% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pCor.R
\name{pCor}
\alias{pCor}
\title{This function aims to construct one more flexible pcor()
All X, Y C should be numeric}
\usage{
pCor(X, Y, C, cor_method = c("pearson", "kendall", "spearman"), allnumeric = T)
}
\arguments{
\item{X}{A data frame of X variables (All variables needs to be numeric)}

\item{Y}{A dataframe ofY variables (needs to be numeric)}

\item{C}{A data frame of covariates (All variables needs to be numeric)}

\item{cor_method}{from cor()}

\item{allnumeric}{don't know why it has problem logical if True all variables are transformed to numeric}
}
\value{
A vector with adjusted partial correlations


install.packages(remotes)
library(remotes)
install_gitlab('CarlBrunius/MUVR@MUVR2')
library(MUVR)
X=Xotu[,1:4]
Y=as.numeric(Yotu)
C=Xotu[,5:7]
}
\description{
This function aims to construct one more flexible pcor()
All X, Y C should be numeric
}
