% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevel_outcome.R
\name{relevel_outcome}
\alias{relevel_outcome}
\title{relevel the outcome matrix}
\usage{
relevel_outcome(outcome, REF)
}
\arguments{
\item{outcome}{In the format of dataframe, could be binary, categorical (>2 groups) and numeric. It is recommeneded that you put outcome one at a time}

\item{REF}{A vector of which variabel is selected as reference}

\item{TPObject}{A TriPlotObject}
}
\value{
A list of risk estimation matrix with components in rows and (estimates, margin-of-error and p-values) in columns. The margin-of-error corresponds to half the width of the confidence interval (i.e. z * se).
}
\description{
relevel the outcome matrix
}
\examples{
See example under triPlot(
}
