% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_stuff.R
\name{removestuff}
\alias{removestuff}
\title{Remove correlations from TriPlotObject}
\usage{
removestuff(TPObject, remove_corr = NULL, remove_risk = NULL)
}
\arguments{
\item{TPObject}{A TriPlotObject}

\item{remove_corr}{Which correlation(s) to remove (either 'all', a number or a vector with names)}

\item{remove_risk}{Which correlation(s) to remove (either 'all', a number or a vector with names)}
}
\value{
A TriPlotObject with added risk estimates
}
\description{
Remove correlations from TriPlotObject
}
\examples{
See example under triPlot()
}
