% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_around_xory.R
\name{rotate_around_xory}
\alias{rotate_around_xory}
\title{rotate around the center, a specific x row or specific y  columns}
\usage{
rotate_around_xory(Dataframe, rotate_around_x = NULL, rotate_around_y = NULL)
}
\arguments{
\item{Dataframe}{a dataframe}

\item{rotate_around_x}{one x column that the matrix will rotate around, coud be a number or column names}

\item{rotate_around_y}{one y column that the matrix will rotate around, coud be a number or column names}
}
\value{
Detailed information and heatmap of correlations and risks in TriPlotObject
}
\description{
rotate around the center, a specific x row or specific y  columns
}
