% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Casualgraph.R
\name{Showdiagram}
\alias{Showdiagram}
\title{Casual diagram}
\usage{
Showdiagram(
  exposure = NULL,
  outcome = NULL,
  mediator = NULL,
  covariate_EM = NULL,
  covariate_EO = NULL,
  covariate_MO = NULL,
  effect_modifier = NULL
)
}
\arguments{
\item{exposure}{A character vector or NULL}

\item{outcome}{A character vector or NULL}

\item{mediator}{These could be metabolomics data (How to map it in detail will be developed)}

\item{covariate_EM}{A character vector or NULL}

\item{effect_modifier}{A character vector or NULL}

\item{covariate_EC}{A character vector or NULL}

\item{covariate_MC}{A character vector or NULL}
}
\value{
A TriPlotObject
}
\description{
Casual diagram
}
