% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriplotGUI.R
\name{TriplotGUI}
\alias{TriplotGUI}
\title{Produce a triplot with loadings, correlations and risk estimates}
\usage{
TriplotGUI(
  TPObject,
  first_PC = 1,
  second_PC = 2,
  plotLoads = TRUE,
  plotScores = FALSE,
  plotCorr = TRUE,
  plotRisk = TRUE,
  loadLabels = TRUE,
  loadArrowLength = 0.02,
  loadCut = 0,
  loadLim,
  colCorr,
  pchCorr = 16,
  whichCorr = NULL,
  corLim,
  colRisk,
  pchRisk = 15,
  whichRisk = NULL,
  riskLim,
  riskWhisker_percentage = 0.1,
  riskOR = T,
  scoreLabels = FALSE
)
}
\arguments{
\item{TPObject}{A TriPlotObject}

\item{first_PC}{The first PC to plot}

\item{second_PC}{The second PC to plot}

\item{plotLoads}{Whether to plot loadings (TRUE; default) or suppress them (FALSE)}

\item{plotScores}{Whether to plot scores (TRUE) or suppress them (FALSE; default)}

\item{plotCorr}{Whether to plot correlations (TRUE; default) or suppress them (FALSE)}

\item{plotRisk}{Whether to plot risk estimates (TRUE; default) or suppress them (FALSE)}

\item{loadLabels}{Whether to plot variable loading labels (TRUE; default) or not (FALSE)}

\item{loadArrowLength}{Length of arrow tip (in inces)}

\item{loadCut}{Loadings below the cut are plotted in light grey and without labels}

\item{loadLim}{Plot range for loadings}

\item{colCorr}{Color vector for correlations}

\item{pchCorr}{Plotting character for correlations}

\item{whichCorr}{Which correlations to plot (vector of numbers)}

\item{corLim}{Plot range for correlations}

\item{colRisk}{Color vector for risk estimates}

\item{pchRisk}{Plotting character for risk estimates}

\item{whichRisk}{Which risk estimates to plot (vector of numbers)}

\item{riskLim}{Plot range for risks}

\item{riskWhisker_percentage}{default is 1/10 of the confidence interval}

\item{riskOR}{Specify whether to antilog risk layer scale (useful for log:ed risk estimates)}

\item{scoreLabels}{Whether to plot observation score labels (TRUE) or not (FALSE; default)}

\item{size}{font size of the name of correlation and risk variables}
}
\value{
A triplot obtained using base R plotting.
}
\description{
Produce a triplot with loadings, correlations and risk estimates
}
\examples{
# Make a PCA
pca <- prcomp(myData, scale.=TRUE, center=TRUE)
# Prepare a TriPlotObject (TPO)
tpo <- makeTPO(scores=pca$x, loadings=pca$rotation, compLimit=10)
# Make a correlation matrix with TPO scores
# NB: This step can be achieved in other ways, e.g. by making partial correlations.
corMat <- makeCorr(tpo, corrData)
# Add correlations to TPO
tpo <- addCorr(tpo, corMat) # More correlations can be added to TPO by further calls to addCorr
# Make risk modelling
# NB: This step can be achieved in other ways, e.g. by standard logistic correlations or by adjusting for other factors/covariates.
risk <- crudeCLR(tpo, caseStat, pairs)
# Add risk estimates to TPO
tpo <- addRisk(tpo,risk,'crudeOR')
# Make a heatmap to check which components may be most interesting to examine
checkTPO(tpo)
# Produce a triplot
triPlot(tpo) # Standard triplot for components 1 and 2.
triPlot(tpo, comps=c(3,5)) # Standard triplot for components 3 and 5.
# See tutorial at triplot repository for more examples
}
