% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_best_transformation.R
\name{which_best_transformation}
\alias{which_best_transformation}
\title{Scan the data frame by each column and see which one is. transformation and}
\usage{
which_best_transformation(Dataframe, Group = NULL, p_threshold = 0.05)
}
\arguments{
\item{Dataframe}{rows as observations columns as}

\item{Group}{Group vairable if you want to separate one variable to different groups to examine normality}

\item{p_threshold}{a p value, usually set as 0.05}
}
\value{
best trnsformation type
}
\description{
Scan the data frame by each column and see which one is. transformation and
}
